refactor: conpound pattern을 사용하여 중복된 로직 제거 (login)
기존 약간의 스타일과 동작의 차이로 인한 불필요한 컴포넌트 생성을 conpound pattern을 적용하여 한 파일 내에서 관리하도록 변경

#31
