--- FE/src/components/login/slack/LoginSection.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-interface LoginSectionProps {
-  title: string;
-  loginBtnComponent: React.ReactNode;
-}
-
-const LoginSection = ({ title, loginBtnComponent }: LoginSectionProps) => {
-  return (
-    <div className="flex flex-col items-center gap-6">
-      <p className="font-light">{title}</p>
-      {loginBtnComponent}
-    </div>
-  );
-};
-
-export default LoginSection;
